+++ packages/server/src/__tests__/integration/connectors/postgres-sync-cloud-gate.test.ts
@@ -9,7 +9,7 @@
  */
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import type { Env } from '../../../index';
-import { createSyncRun } from '../../../utils/queue-helpers';
+import { createSyncRun } from '../../../runs/queue-service';
 import { cleanupTestDatabase, getTestDb } from '../../setup/test-db';
 import { post } from '../../setup/test-helpers';
 import {
